Moves the backend message monitoring to a new store
