+++ frontend/app/src/components/app/AppMessages.vue
@@ -8,40 +8,31 @@
       @dismiss="dismissMessage()"
     />
     <startup-error-screen
-      v-if="startupError.length > 0"
-      :message="startupError"
+      v-if="startupErrorMessage.length > 0"
+      :message="startupErrorMessage"
       fatal
     />
-    <mac-os-version-unsupported v-if="macosUnsupported" />
-    <win-version-unsupported v-if="winUnsupported" />
+    <mac-os-version-unsupported v-if="isMacOsVersionUnsupported" />
+    <win-version-unsupported v-if="isWinVersionUnsupported" />
   </fragment>
 </template>
 
 <script setup lang="ts">
+import MessageDialog from '@/components/dialogs/MessageDialog.vue';
+import MacOsVersionUnsupported from '@/components/error/MacOsVersionUnsupported.vue';
+import StartupErrorScreen from '@/components/error/StartupErrorScreen.vue';
+import WinVersionUnsupported from '@/components/error/WinVersionUnsupported.vue';
 import Fragment from '@/components/helper/Fragment';
+import { useBackendMessagesStore } from '@/store/backend-messages';
 import { useMessageStore } from '@/store/message';
 
-const StartupErrorScreen = defineAsyncComponent(
-  () => import('@/components/error/StartupErrorScreen.vue')
-);
-const MessageDialog = defineAsyncComponent(
-  () => import('@/components/dialogs/MessageDialog.vue')
-);
-const MacOsVersionUnsupported = defineAsyncComponent(
-  () => import('@/components/error/MacOsVersionUnsupported.vue')
-);
-const WinVersionUnsupported = defineAsyncComponent(
-  () => import('@/components/error/WinVersionUnsupported.vue')
-);
-
-defineProps({
-  startupError: { required: true, type: String },
-  macosUnsupported: { required: true, type: Boolean },
-  winUnsupported: { required: true, type: Boolean }
-});
-
+const {
+  startupErrorMessage,
+  isMacOsVersionUnsupported,
+  isWinVersionUnsupported
+} = storeToRefs(useBackendMessagesStore());
 const store = useMessageStore();
-const { message } = toRefs(store);
+const { message } = storeToRefs(store);
 const { setMessage } = store;
 const dismissMessage = () => setMessage();
 </script>
