+++ frontend/app/src/router/index.ts
@@ -28,7 +28,7 @@ const routes = setupLayouts([
   },
   {
     path: Routes.USER_CREATE,
-    component: async () => import('../pages/user/login/index.vue'),
+    component: async () => import('../pages/user/create/index.vue'),
     meta: {
       layout: 'login'
     }
